requirements.txt
amqp==5.3.1
asgiref==3.11.1
bandit==1.9.4
billiard==4.2.4
boolean.py==5.0
CacheControl==0.14.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cyclonedx-python-lib==11.7.0
defusedxml==0.7.1
dj-database-url==3.1.2
Django==6.0.4
django-htmx==1.27.0
djangorestframework==3.17.1
et_xmlfile==2.0.0
Faker==40.13.0
filelock==3.25.2
gunicorn==25.3.0
idna==3.11
kombu==5.6.2
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
openpyxl==3.1.5
packageurl-python==0.17.6
packaging==26.0
pillow==12.2.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.6
prompt_toolkit==3.0.52
psycopg==3.3.3
psycopg-binary==3.3.3
py-serializable==2.1.0
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
requests==2.33.1
rich==15.0.0
ruff==0.15.10
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.5
stevedore==5.7.0
tomli==2.4.1
tomli_w==1.2.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.12.0
